[Semaphore] fix issue #20

+++ lib/semaphore/StallardOSsem.hpp
@@ -6,7 +6,7 @@
 
 class StallardOSSemaphore
 {
-    volatile uint16_t val = 1;
+    volatile uint32_t val = 1;
 public:
     StallardOSSemaphore();
